--- BANDITS/Test_env/example_code_john.py
@@ -1,8 +1,0 @@
-# example of changeble output names
-# basename = 'out{}.tif'
-# tuning = 3
-# outname = os.path.join(raster_folder, basename.format(tuning))
-
-# example of new sub directory creation
-# subdir = os.path.join(raster_folder, 'sub')
-# os.makedirs(subdir, exist_ok=True)
