--- BANDITS/Test_env/mask_mean_test.py
@@ -1,41 +1,0 @@
-import numpy as np
-import matplotlib.pyplot as plt
-# x = [-15.885208, -15.724949, -15.784478, -15.934047, -16.301727, -16.065033, -16.590078, -15.364529, -15.47003, -15.418651, -15.598436, -14.6828165, -15.622087, -15.391632, -13.56217, -15.697519, -23.17588, -16.352894, -24.73628, -28.51331, -16.401258, -18.21114, -19.96806, -28.745222, -18.254887, -21.03671, -19.708326, -16.226046, -16.43482, -16.639982, -16.562746, -16.359783, -16.417582, -16.520609, -16.668991, -16.523668, -16.779423, -15.54529, -16.59029, -16.604824, -16.512249, -16.812452, -17.021282, -17.055443, -17.170143, -16.936829, -16.734245, -16.267109, -16.776241, -16.462193, -16.068403, -16.42881, -14.762305, -16.328651, -16.156425, -15.337967, -16.485165, -16.781418, -16.609539, -16.664145, -13.92276, -16.585052, -16.79633, -16.649246, -16.788126, -16.786345, -16.916979, -16.619366, -16.838255, -16.867512, -16.87262, -16.706463, -16.766535, -16.76477, -16.90309, -16.867535, -16.885073, -16.78429, -16.746696, -13.035741, -16.854515, -16.655884, -16.48683, -16.957172, -16.557966, -16.226635, -16.526676, -16.339426, -16.557194, -16.279974, -16.852638, -16.695269, -16.457104, -16.635595, -16.849016, -16.646025, -16.475525, -14.269246, -16.422726, -16.648731, -13.1904955, -16.497232, -16.279396, -15.549868, -16.47346, -16.509253, -16.342031, -16.737616, -16.849426, -16.748875, -16.659525, -16.721588, -16.221458, -16.694113, -16.475075, -16.543999, -16.758076]
-# y = np.convolve([-15.885208, -15.724949, -15.784478, -15.934047, -16.301727, -16.065033, -16.590078, -15.364529, -15.47003, -15.418651, -15.598436, -14.6828165, -15.622087, -15.391632, -13.56217, -15.697519, -23.17588, -16.352894, -24.73628, -28.51331, -16.401258, -18.21114, -19.96806, -28.745222, -18.254887, -21.03671, -19.708326, -16.226046, -16.43482, -16.639982, -16.562746, -16.359783, -16.417582, -16.520609, -16.668991, -16.523668, -16.779423, -15.54529, -16.59029, -16.604824, -16.512249, -16.812452, -17.021282, -17.055443, -17.170143, -16.936829, -16.734245, -16.267109, -16.776241, -16.462193, -16.068403, -16.42881, -14.762305, -16.328651, -16.156425, -15.337967, -16.485165, -16.781418, -16.609539, -16.664145, -13.92276, -16.585052, -16.79633, -16.649246, -16.788126, -16.786345, -16.916979, -16.619366, -16.838255, -16.867512, -16.87262, -16.706463, -16.766535, -16.76477, -16.90309, -16.867535, -16.885073, -16.78429, -16.746696, -13.035741, -16.854515, -16.655884, -16.48683, -16.957172, -16.557966, -16.226635, -16.526676, -16.339426, -16.557194, -16.279974, -16.852638, -16.695269, -16.457104, -16.635595, -16.849016, -16.646025, -16.475525, -14.269246, -16.422726, -16.648731, -13.1904955, -16.497232, -16.279396, -15.549868, -16.47346, -16.509253, -16.342031, -16.737616, -16.849426, -16.748875, -16.659525, -16.721588, -16.221458, -16.694113, -16.475075, -16.543999, -16.758076], (1/9.0)*np.ones((9)), "valid")
-# print((1/3.0)*np.ones((3)))
-# #print(np.convolve([1,2,3],[0,,0.5], 'valid'))
-#
-# plt.plot(x)
-# plt.show()
-#
-# plt.plot(y)
-# plt.show()
-
-test = [-14.5648756 -15.07782364 -15.62544441 -16.15462303 -16.28585815
- -16.28585815 -16.28585815 -16.28585815 -16.41970634 -16.41970634
- -16.41970634 -16.41970634 -16.53263855 -16.67783737 -16.53263855
- -16.41970634 -16.53263855 -16.53263855 -16.8511734  -16.53263855
- -16.53263855 -16.53263855 -16.8511734 ]
-#test2 = [0 1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 19 20 21 22]
-
-#slope, intercept, r_value, p_value, std_err = stats.linregress(test,test2)
-
-#print(slope)
-
-mini = min(test)
-print(mini)
-
-
-l = [-1,-2,-3,-4,-5,-6,-7,-8,-9]
-
-print(np.percentile(l, [10,90]))
-
-print(np.percentile(l, 80))
-
-print(np.percentile(l, 10))
-
-print(np.mean(l))
-print(np.percentile(l, 50))
-
-print(np.percentile(l, 100))
-print(np.percentile(l, 0))
