+++ BanDiTS/Test_env/example_old.py
@@ -1,4 +1,4 @@
-from BANDITS import preprocessing, apply_along_axis, export_arr, statistical_functions, filter_functions
+from BanDiTS import preprocessing, apply_along_axis, export_arr, statistical_functions, filter_functions
 from pathos import multiprocessing as mp
 from datetime import datetime
 import numpy as np
