+++ BanDiTS/example.py
@@ -1,6 +1,6 @@
-from BANDITS import preprocessing, apply_along_axis, export_arr, filter_functions, statistical_functions
-from BANDITS.statistical_functions import *
-from BANDITS.filter_functions import *
+from BanDiTS import preprocessing, apply_along_axis, export_arr, filter_functions, statistical_functions
+from BanDiTS.statistical_functions import *
+from BanDiTS.filter_functions import *
 from pathos import multiprocessing as mp
 from datetime import datetime
 import numpy as np
